requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-cloud-storage
python-dotenv

